pyproject.toml
[tool.poetry]
name = "pymacio"
version = "0.1.0"
description = ""
authors = ["Anthony Dawson <anthonypdawson@gmail.com>"]
readme = "README.md"
repository = ""
packages = [{ include = "pymacio" }]


[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
arg-parser = "^1.0.5"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

